Fixes tests to refer current image
