+++ tests/integration/test_docker_image.py
@@ -20,16 +20,6 @@ import pytest
 
 
 @pytest.fixture
-def docker_client() -> docker.DockerClient:
-    return docker.from_env()
-
-@pytest.fixture
-def docker_image_key(docker_client: docker.DockerClient) -> str:
-    image_key = "simcore/services/comp/osparc-opencor:latest"
-    docker_images = [image for image in docker_client.images.list() if any(image_key in tag for tag in image.tags)]
-    return docker_images[0].tags[0]
-
-@pytest.fixture
 def docker_image(docker_client: docker.DockerClient, docker_image_key: str) -> docker.models.images.Image:
     docker_image = docker_client.images.get(docker_image_key)
     assert docker_image
